feat(tests): enhance Docker setup for Fake GCS Server and add blob operations tests
