--- tests/test_runner.py
@@ -1,4 +1,0 @@
-import pytest
-from pathlib import Path
-
-pytest.main([(Path(__file__).parent / "tests.py").as_posix()])
